Changed the brute forced "obj" converter and moved lambda-ike functions into utils.py
